Modified constructor to initialize holders of incoming variables.
Added a method setAccel() that accepts a pointer to an array of integers
that holds a per-peer-process counter of elements that are potentially
going to overlap with this process's elements. This method is to accelerate
the whole procedure of finding overlaps by knowing "a priori" which of the
peer processes to exclude from tests.

Added a method prepare() to perform operations that prepare the object to
perform matching. Pointers and an internal size have also been added to the
private members of the object. The method performs sanity checks before it
proceeds; if the object has not been provided external data it returns 1.
Internal arrays are allocated for performing peer-process element exchanges
and arrays are built to hold peer process (including own) data. (There's the
potential to severely bloat the memory in the event of full or large overlap
of process elements.) The method performs a linear-in-p broadcast of the
element data, which is a potential performance bottleneck; each process is
to keep the data it needs given the acceleration array counts provided.

Modified the deconstructor to clean up all arrays attached to the object's
pointers.
IN 2017/04/25