Corrected error on array index displacement during MPI_Bcast of surface 2. (This was found in the process of building the present step in the code.)
Added a vector of lists of overlap structures to the object. The array is
meant to be sized/indexed by the local owned surface 1 elements. Each of
the lists holds the owing process number, the element index, and the area
of overlap with each of the elements of surface 2. At present, there is a
test devised to check for memory leaks that is bloating the arrays with
a different kind of data; this will be deleted in the next commit.
IN 2017/04/25