Modified the object's perform() method to perform matching. In order to do this, the method prepare() that builds the contents of the idata[] array via per-process broadcasts was modified to include the correct integer type data that is needed; the contents of this array may change at the highest level as needed in the future.
In the perform() method, a double loop of N*M overlap queries is made, and
this is NOT a very efficient way to perform ths search for overlaps. It is
relying on the "acceleration array" to trim the search space ahead of time.
This algorithm may have to change in the future; it was coded here because
of the legacy operations that sparked the creation of this functionality in
the first place.

To find the individual element(1) to element(2) overlap, a function that
is built by the user must be employed. The object is given a pointer to the
function that is to be used; this is done with the setter method that is
setOverlapFunction(). In order for this function to be external from the
object (and library structure), a global pointer variable is introduced on
the object's global space side called incg_FaceOverlapFunction. This is a
pointer to a function that accepts two sized arrays of node coefficients.
The global variable pointer is to be set externally from a C code such that
the library can use it when performing face-matching.
IN 2017/04/26