Corrected the error-trapping handling in the standalone driver routine that is to be used for testing (as discussed in the previous commit).
Created the incg_Facematch_Init() and incg_Facematch_Term() functions that
operate on face-matching object handles. The functionality now allows for
multiple, simultaneous, objects to be active, each associated with a handle.
A C++ vector container holds pointers to these objects corresponding to the
handles; the handles are the index in the vector array. This functionality
was tested (in a loop for memory leaks with interleaved handles) and works
as expected.
IN 2017/04/27